requirements.txt
Django==4.0.3
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
executing==0.8.3
ipdb==0.13.9
ipython==8.2.0
Markdown==3.3.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
psycopg2==2.9.3
Pygments==2.11.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
django-filter==21.1
django-rest-swagger==2.2.0
